requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
asyncpg==0.30.0
cfgv==3.4.0
click==8.2.1
dill==0.4.0
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.14
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
identify==2.6.12
idna==3.10
isort==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
pydantic==1.10.22
pylint==3.3.7
pylint-plugin-utils==0.9.0
pylint-pydantic==0.3.5
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
virtualenv==20.31.2
